Add MDC context to logs
This change will add several MDC context parameters to the log
statement, to provide additional information in a structured maner
for the log output.

Added the ok flag to the HealthCheckReportDto as an additional boolean
for if the check passed or not, that is used in the logger.

Created a small MDCStack utility inside the default logger, that is
useful when putting several variables on the MDC for a short time. In
this case, for a single log statement.

Change-Id: Iff687659184272499928f401271d8df7cc5bdeb0